+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0] - 2026-03-13
+
+### Added
+
+- `Util.Text` with helpful text utilities
+- `StringExtensions` with helpful string formatting extensions
+
 ## [1.10.0] - 2026-03-12
 
 ### Changed